+++ game/missiongenerator/aircraft/waypoints/baiingress.py
@@ -10,6 +10,7 @@ from .pydcswaypointbuilder import PydcsWaypointBuilder
 
 class BaiIngressBuilder(PydcsWaypointBuilder):
     def add_tasks(self, waypoint: MovingPoint) -> None:
+        self.register_special_ingress_points()
         waypoint.tasks.append(OptFormation.trail_open())
         # TODO: Add common "UnitGroupTarget" base type.
         group_names = []