+++ game/missiongenerator/aircraft/waypoints/deadingress.py
@@ -10,7 +10,8 @@ from .pydcswaypointbuilder import PydcsWaypointBuilder
 
 class DeadIngressBuilder(PydcsWaypointBuilder):
     def add_tasks(self, waypoint: MovingPoint) -> None:
-        self.register_special_waypoints(self.waypoint.targets)
+        self.register_special_strike_points(self.waypoint.targets)
+        self.register_special_ingress_points()
 
         target = self.package.target
         if not isinstance(target, TheaterGroundObject):